+++ packages/media-parser/src/boxes/iso-base-media/process-box.ts
@@ -1,8 +1,8 @@
 import type {BufferIterator} from '../../buffer-iterator';
 import {Log} from '../../log';
 import {registerTrack} from '../../register-track';
-import type {BoxAndNext} from '../../run-parse-iteration';
 import type {ParserState} from '../../state/parser-state';
+import type {BoxAndNext} from './base-media-box';
 import {parseEsds} from './esds/esds';
 import {parseFtyp} from './ftyp';
 import {getIsoBaseMediaChildren} from './get-children';
