+++ packages/media-parser/src/metadata/metadata-from-iso.ts
@@ -1,10 +1,11 @@
+import type {RegularBox} from '../boxes/iso-base-media/base-media-box';
 import type {IlstBox} from '../boxes/iso-base-media/meta/ilst';
 import {
 	getMoovBox,
 	getTkhdBox,
 	getTraks,
 } from '../boxes/iso-base-media/traversal';
-import type {IsoBaseMediaStructure, RegularBox} from '../parse-result';
+import type {IsoBaseMediaStructure} from '../parse-result';
 import {truthy} from '../truthy';
 import type {MetadataEntry} from './get-metadata';
