+++ packages/media-parser/src/run-parse-iteration.ts
@@ -6,13 +6,8 @@ import {parseWav} from './boxes/wav/parse-wav';
 import {parseWebm} from './boxes/webm/parse-webm-header';
 import type {BufferIterator} from './buffer-iterator';
 import {initVideo} from './init-video';
-import type {IsoBaseMediaBox, ParseResult} from './parse-result';
+import type {ParseResult} from './parse-result';
 import type {ParserState} from './state/parser-state';
-
-export type BoxAndNext = {
-	box: IsoBaseMediaBox | null;
-	skipTo: number | null;
-};
 
 export const runParseIteration = async ({
 	iterator,
