+++ packages/media-parser/src/samples-from-moof.ts
@@ -1,10 +1,11 @@
+import type {IsoBaseMediaBox} from './boxes/iso-base-media/base-media-box';
 import {
 	getTfdtBox,
 	getTfhdBox,
 	getTrunBoxes,
 } from './boxes/iso-base-media/traversal';
 import type {SamplePosition} from './get-sample-positions';
-import type {AnySegment, IsoBaseMediaBox} from './parse-result';
+import type {AnySegment} from './parse-result';
 
 const getSamplesFromTraf = (
 	trafSegment: IsoBaseMediaBox,
